+++ adapter/bidmachine/CHANGELOG.md
@@ -7,6 +7,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 📋 [Release Notes](https://developers.bidmachine.io/sdk/general/android/android-changelog)
 
 ## [Unreleased]
+### Added
+- Relayed ML floor predictions from the bid response custom params into the ad unit ext reported in stats and auction info
 
 ## [3.7.1.0] - 2026-06-01
 ### Changed